feat: 보관함 조회 서버 연결

+++ app/src/main/java/com/grusie/sharingmap/ui/model/StorageUiModel.kt
@@ -3,6 +3,6 @@ package com.grusie.sharingmap.ui.model
 data class StorageUiModel(
     val id: Long = 0,
     val title: String = "",
-    val count: Long = 0,
+    val count: Int = 0,
     val feeds: List<ArchiveUiModel> = emptyList(),
 )
